abode/requirements.txt
fastapi
uvicorn
motor
python-dotenv
python-multipart
pymupdf
python-docx
pymongo
numpy
scikit-learn
transformers
torch
jsonschema